+++ treeshr/TreeGetDbi.c
@@ -11,7 +11,7 @@ extern void *DBID;
 #define min(a,b) (((a) < (b)) ? (a) : (b))
 #endif
 int TreeGetDbi(struct dbi_itm *itmlst) {return _TreeGetDbi(DBID,itmlst);}
-#define set_retlen(length) if (lst->buffer_length < length) { status = TreeBUFFEROVF; break; } else retlen=length
+#define set_retlen(length) if ((unsigned int)lst->buffer_length < length) { status = TreeBUFFEROVF; break; } else retlen=length
 #define CheckOpen(db) if (!db || !db->open) {status=TreeNOT_OPEN;break;}
 #define set_ret_char(val) memset(lst->pointer, 0, lst->buffer_length); *((char *)lst->pointer) = val
